Final working version of courseWork for APS

+++ Sem5/SoftSysArch/src/additionalSystem/inits.cpp
@@ -1,6 +1,6 @@
 #include "inits.hpp"
 
-std::vector< source > initSources(size_t numOfSrc, const std::vector< double > & lambda, double delay, printer * print)
+std::vector< source > initSources(size_t numOfSrc, const std::vector< long double > & lambda, long double delay, printer * print)
 {
   std::vector< source > src;
   for (size_t i = 0; i < numOfSrc; ++i)
@@ -20,7 +20,7 @@ std::vector< buffer > initBuffers(size_t numOfBuffs, size_t buffSize, printer * 
   return buff;
 }
 
-std::vector< handler > initHandlers(size_t numOfHands, double delay, double handWork, printer * print)
+std::vector< handler > initHandlers(size_t numOfHands, long double delay, long double handWork, printer * print)
 {
   std::vector< handler > handlers;
   for (size_t i = 0; i < numOfHands; ++i)
@@ -30,7 +30,7 @@ std::vector< handler > initHandlers(size_t numOfHands, double delay, double hand
   return handlers;
 }
 
-std::vector< sourceBufferDispatcher > initFDisps(size_t numOfFDisps, std::vector< source > * sources, std::vector< buffer > * buff, double delay)
+std::vector< sourceBufferDispatcher > initFDisps(size_t numOfFDisps, std::vector< source > * sources, std::vector< buffer > * buff, long double delay)
 {
   std::vector< sourceBufferDispatcher > firstDisp;
   for (size_t i = 0; i < numOfFDisps; ++i)
@@ -40,7 +40,7 @@ std::vector< sourceBufferDispatcher > initFDisps(size_t numOfFDisps, std::vector
   return firstDisp;
 }
 
-std::vector< bufferHandlerDispatcher > initSDisps(size_t numOfSDisps, std::vector< handler > * handlers, std::vector< buffer > * buff, double delay)
+std::vector< bufferHandlerDispatcher > initSDisps(size_t numOfSDisps, std::vector< handler > * handlers, std::vector< buffer > * buff, long double delay)
 {
   std::vector< bufferHandlerDispatcher > secondDisp;
   for (size_t i = 0; i < numOfSDisps; ++i)
