+++ Sem5/SoftSysArch/src/additionalSystem/inits.hpp
@@ -11,10 +11,10 @@
 #include "../mainSystem/source.hpp"
 #include "../mainSystem/sourceBufferDispatcher.hpp"
 
-std::vector< source > initSources(size_t numOfSrc, const std::vector< double > & lambda, double delay, printer * print);
+std::vector< source > initSources(size_t numOfSrc, const std::vector< long double > & lambda, long double delay, printer * print);
 std::vector< buffer > initBuffers(size_t numOfBuffs, size_t buffSize, printer * print);
-std::vector< handler > initHandlers(size_t numOfHands, double delay, double handWork, printer * print);
-std::vector< sourceBufferDispatcher > initFDisps(size_t numOfFDisps, std::vector< source > * sources, std::vector< buffer > * buff, double delay);
-std::vector< bufferHandlerDispatcher > initSDisps(size_t numOfSDisps, std::vector< handler > * handlers, std::vector< buffer > * buff, double delay);
+std::vector< handler > initHandlers(size_t numOfHands, long double delay, long double handWork, printer * print);
+std::vector< sourceBufferDispatcher > initFDisps(size_t numOfFDisps, std::vector< source > * sources, std::vector< buffer > * buff, long double delay);
+std::vector< bufferHandlerDispatcher > initSDisps(size_t numOfSDisps, std::vector< handler > * handlers, std::vector< buffer > * buff, long double delay);
 
 #endif
