+++ Sem5/SoftSysArch/src/additionalSystem/printer.hpp
@@ -11,15 +11,15 @@
 struct sourceStats
 {
   size_t generates_, deleted_, sucsessfull_;
-  double timeInSystem_, workingTime_;
+  long double timeInSystem_, workingTime_;
   std::string name_;
-  std::vector < double > appsTimes_;
+  std::vector < long double > appsTimes_;
 };
 
 struct handlerStats
 {
   size_t sucsessfull_;
-  double workingTime_;
+  long double workingTime_;
   bool isWorking_;
 };
 
@@ -35,16 +35,16 @@ class printer
     printer(size_t numOfSrc, const std::vector< std::string > & sourcesNames, size_t numOfBuffs, size_t numOfHands, size_t buffSize, std::ostream * out);
 
     void setMod(bool autoMod);
-    void printStep(size_t step, double totalTime);
-    void printRes(double totalTime);
-    void endSource(size_t id, double fullTime, double workTime);
+    void printStep(size_t step, long double totalTime);
+    void printRes(long double totalTime);
+    void endSource(size_t id, long double fullTime, long double workTime);
     void printSource(size_t id, size_t actualAppId, size_t priority);
     void printBufferIn(size_t id, application app);
     void printBufferOut(size_t id);
     void printBufferDel(size_t id, application app);
     void printHandlerIn(size_t id, application app);
     void printHandlerOut(size_t id, application app);
-    void endHandler(size_t id, double workTime);
+    void endHandler(size_t id, long double workTime);
 
   private:
     size_t buffSize_;
